first round of refactoring to move cache indexers into separate file
