+++ dumpcache.c
@@ -310,8 +310,9 @@ static const struct file_operations dumpcache_ops = {
 // register or read the data side L1 or L2 array contents into the
 // DL1DATAn register.
 //
-static inline void asm_ramindex_msr(u32 ramindex)
+static inline void asm_ramindex_msr(const char *whence, u32 ramindex)
 {
+        (void)whence;
 	asm volatile(
 	    "sys #0, c15, c4, #0, %0\n"
 	    "dsb sy\n"  //        data sync barrier
@@ -362,64 +363,7 @@ static inline void asm_ramindex_insn_mrs(u32 *ildata, u8 sel)
 	}
 }
 
-//
-// Get Tag of L2 cache entry at (index, way).
-// Tag bank select ignored, 2MB L2 cache assumed.
-//
-static inline void get_L2_tag(u32 index, u32 way, u32 *dl1data)
-{
-	u32 ramid    = 0x10;  // L2 Tag RAM magic number (page 4-184)
-	u32 ramindex = (ramid << 24) + (way << 18) + (index << 6);
-
-	asm_ramindex_msr(ramindex);
-	asm_ramindex_data_mrs(dl1data, 0x01);  // reads just dl1data[0]
-
-	// Check if MOESI state is invalid, and if so, zero out the address
-	if (((*dl1data) & 0x03UL) == 0) {
-          *dl1data = 0;
-          return;
-	}
-	// Isolate the tag
-	*dl1data &= ~(0x03UL);
-	*dl1data <<= 12;
-	*dl1data |= (index << 5);
-}
-
-#define MASK2(ub, lb) ((0x1UL<<((ub)-(lb)+1)) - 1)
-
-static inline void get_L1Itag(u32 index, u32 way, uint32_t *instructions) {
-  //
-  // index is the virtual address. See figure 4-58
-  //
-  u32 ramid    = 0x00;  // L1-I Tag RAM
-  u32 ramindex = 0
-    | (ramid << 24)
-    | ((way & 0x3) << 18)
-    | ((index & MASK2(13,6)) << 5);
-  instructions[0] = 0;
-  instructions[1] = 0;
-  asm_ramindex_msr(ramindex);
-  asm_ramindex_insn_mrs(instructions, 0x03);  // get insn[0] and insn[1]
-  // instructions[0] = 0xbadf00dUL;  // dummy physical address
-  // instructions[1] = 0x1111111UL;  // dummy 2-bit field: valid and non-secure id
-}
-
-static inline void get_L1Iinsn(u32 index, u32 way, u32 *instructions)
-{
-  //
-  // index is the virtual address. See figure 4-58
-  //
-  u32 ramid    = 0x01;  // L1-I Data RAM
-  u32 ramindex = 0
-    | (ramid << 24)
-    | ((way & 0x3) << 18)
-    | ((index & MASK2(13,3)) << 3)
-    ;
-  instructions[0] = 0;
-  instructions[1] = 0;
-  asm_ramindex_msr(ramindex);
-  asm_ramindex_insn_mrs(instructions, 0x03);  // get insn[0] and insn[1]
-}
+#include "cache_operations.c"
 
 bool rmap_one_func(struct page *page, struct vm_area_struct *vma, unsigned long addr, void *arg)
 {
@@ -601,42 +545,6 @@ static int dump_all_indices(void) {
 		}
 	}
 	return 0;
-}
-
-static int get_Cortex_L1_Tag(void) {
-    uint32_t way, bank, set;
-    struct Cortex_L1_I_Tag_Cache *cache =
-        (struct Cortex_L1_I_Tag_Cache *)cur_sample;
-    for (way = 0; way < 3; way++) {
-        for (bank = 0; bank < 2; bank++) {
-            for (set = 0; set < 128; set++) {
-                struct Cortex_L1_I_Tag_Bank_Line *p =
-                  &cache->way[way].bank[bank].set[set];
-                uint32_t index = (set << 7) | (bank << 6);
-                get_L1Itag(index, way, p->u.instruction);
-            }
-        }
-    }
-    return 0;
-}
-
-static int get_Cortex_L1_Insn(void) {
-    uint32_t way, bank, set, pair;
-    struct Cortex_L1_I_Insn_Cache *cache =
-        (struct Cortex_L1_I_Insn_Cache *)cur_sample;
-    for (way = 0; way < 3; way++) {
-        for (bank = 0; bank < 4; bank++) {
-            for (set = 0; set < 256; set++) {
-                for (pair = 0; pair < 2; pair++) {
-                    struct Cortex_L1_I_Insn_Pair *p =
-                       &cache->way[way].bank[bank].set[set].pair[pair];
-                    uint32_t index = (set << 6) | (bank << 4) | (pair << 3);
-                    get_L1Iinsn(index, way, p->instruction);
-                }
-            }
-        }
-    }
-    return 0;
 }
 
 /* ProcFS interface definition */
